Added another parameter to pass user data to APIs
There are cases when creating a copy or destroying one might take some
external data. One example of this is when I'm creating an array of
vertex buffers in my game engine. Each vertex buffer requires Device
handle to be destroyed. The ususal API makes it hard to pass such
information at any point of time, so I added another parameter to pass
that info.

Now, for further containers and their APIs as well, I'll keep this
parameter, who knows what type user is using.

For me, it'll ease the destruction of array of vertex buffers.

I've also started implementing hash maps in order to compare performance
of game engine when using Vectors and HashMaps in certain scenarios.

Signed-off-by: Siddharth Mishra <admin@brightprogrammer.in>