Added Implemnetation of DenseMap.
DenseMap is a generic container to store key-value pairs.
Provides constant time insertion/deletion/search etc...
DenseMap stores all items with same hash value in clusters
instead of storing in buckets.

Clusters are to DenseMap(s), what Buckets are to SparseMap(s).

DenseMap uses Robin Hood hashing, a modified version of linear
probing to make this possible. The clusters are automatically created
and not hardoded into the implementation.

This commit does not add tests, it needs to be added for multiple other
containers as well, and be improved for remaining ones.

Signed-off-by: Siddharth Mishra <admin@brightprogrammer.in>