Added Tests For BitVector & Block Allocators
This commit restructures the way unit tests are written. This creates
immediate demand to update and write tests for other containers as well
in AnvieUtils. Otherwise, it would be too late and might be a huge
barrier that would never be tackled with.

It took me 4 days just to write these unit tests. The test output is too
noisy at this time, but that is least of my concerns at the moment. I
need to improve that later. To a more clear output of test results.

Also, many tests are failing at the moment. The tests are very
comprehensive and granular but I don't take guarantee of how exhaustive
they are. I hope they are but cannot guarantee this at the moment.

---

The commit also adds two BlockAllocators. One is implemented, other
needs to be implemented in future. It's my guess that it'll take another
two weeks to improve and write new tests for existing containers. I've
postponed this task for very long time now. It's time I deal with it.

Signed-off-by: Siddharth Mishra <admin@brightprogrammer.in>