Add Documentation For SparseMap & Fix Build
This commit adds documentation for SparseMap. I'm too tired now to work
further and write complete documentation for DenseMap as well. It's 1:57
AM IST at the time of writing this.

The SparseMap documentation was also written in a hurry, that needs to
be improved as well.

Happy New Year! :D

Signed-off-by: Siddharth Mishra <admin@brightprogrammer.in>